requirements.txt
torch>=2.0.0
transformers>=4.30.0
flask>=2.0.0
kagglehub>=0.2.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.0.0
pillow>=9.0.0
pytesseract>=0.3.0
python-dotenv>=0.19.0 